+++ adminer/variables.inc.php
@@ -4,7 +4,7 @@ namespace Adminer;
 $status = isset($_GET["status"]);
 page_header($status ? lang('Status') : lang('Variables'));
 
-$variables = ($status ? show_status() : show_variables());
+$variables = ($status ? adminer()->showStatus() : adminer()->showVariables());
 if (!$variables) {
 	echo "<p class='message'>" . lang('No rows.') . "\n";
 } else {